+++ soya/HISTORY.md
@@ -1,5 +1,83 @@
 # Version 0.0.x
 
+## 0.0.75
+
+- Fix `CookieJar.remove()` and `Cookie.createRemoval()` bug, made `secure` a
+  required parameter in removing cookie.
+
+## 0.0.74
+
+- Update `ServerHttpRequest` and `ClientHttpRequest` to have both `getQuery()`
+  and `getQueryParams()`.
+- Extract query string parser for outside usage.
+
+## 0.0.70
+
+- Upgrade to Babel 6.
+- Soya no longer determines babel configuration for your project. You'll need
+  to configure it yourself. This allows different files to have multiple babel
+  configuration, according to Babel's standard for finding .babelrc.
+- Client replace and client resolve no longer working (since nobody uses them,
+  will update them as soon as there's a need).
+- How to update:
+  - Remove all Babel related lines in your package.json into these:
+        
+        ```
+        "babel-core": "^6.4.5",
+        "babel-loader": "^6.2.1",
+        "babel-preset-es2015": "^6.3.13",
+        "babel-preset-react": "^6.3.13",
+        "babel-preset-stage-2": "^6.18.0",
+        "babel-plugin-transform-export-extensions": "^6.8.0",
+        "babel-plugin-transform-object-assign": "^6.8.0",
+        ```
+        
+  - Still in package.json, add the following Babel configuration (you may
+    modify them if you wish):
+    
+        
+        "babel": {
+            "presets": [
+            "es2015",
+            "react",
+            "stage-2"
+          ],
+          "plugins": [
+            "transform-export-extensions",
+            "transform-object-assign"
+          ]
+        }
+        
+        
+  - Change the contents of `webpack.config.js` into this (you can also copy
+    and paste from integration test):
+    
+        
+        var webpack = require('webpack');
+        var Precompiler = require('soya/lib/precompile/Precompiler').default;
+        var WebpackCompiler = require('soya/lib/compiler/webpack/WebpackCompiler').default;
+        var config = require('./config');
+
+        var precompiler = new Precompiler(config.frameworkConfig);
+        precompiler.precompile();
+
+        var webpackConfig = WebpackCompiler.createServerBuildConfig(webpack, config.frameworkConfig);
+        module.exports = webpackConfig;
+        
+  
+  - Remove `node_modules` folder and do a fresh `npm install`.
+- If you're contributing to soya:
+  - Fetch the updates.
+  - Remove `node_modules` folder and do a fresh `npm install`.
+
+## 0.0.69
+
+- Form submission now doesn't assume that the user wants to disable and
+  re-enable the form on submission.
+- Users can use `Form.lockSubmission()` and `Form.unlockSubmission()` if they
+  would like to lock the form when submitting. User can create buttons that
+  listens to this state so it can auto-disable when form is in submission.
+
 ## 0.0.68
 
 - Remove unnecessary import to redux-devtools.